feat: auth implementation without jwt

+++ backend/routes/urls.py
@@ -1,9 +1,12 @@
 from typing import List
 
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends, status, HTTPException
+
 from schemas.url import Url, UrlCreate
 
+import auth.authConfig as authConfig
 import crud.urlCrud as urlCrud
+from schemas.user import User
 
 routes_url = APIRouter()
 
@@ -14,8 +17,10 @@ def get_urls_by_user(owner: int):
 
 
 @routes_url.post("", response_model=Url, status_code=201)
-def create_url(url: UrlCreate):
-    return urlCrud.create_url(url.key, url.url, 0)
+def create_url(url: UrlCreate, current_user: User = Depends(authConfig.get_current_user)):
+    if not current_user:
+        raise HTTPException(status.HTTP_401_UNAUTHORIZED)
+    return urlCrud.create_url(url.key, url.url, current_user.id)
 
 
 @routes_url.get("/{key}", response_model=Url)
@@ -24,5 +29,9 @@ def get_url(key: str):
 
 
 @routes_url.delete("/{key}")
-def delete_url(key: str):
+def delete_url(key: str, current_user: User = Depends(authConfig.get_current_user)):
+    url = urlCrud.get_url(key)
+    if url and url['owner'] != current_user.id:
+        raise HTTPException(status.HTTP_403_FORBIDDEN)
+
     return urlCrud.delete_url(key)
